fixes && projects seeding

+++ backend/repositories/FileRepository.ts
@@ -2,12 +2,11 @@ import * as path from 'path';
 import * as fs from 'fs';
 import { createHash } from 'crypto';
 import * as _ from 'lodash';
-import {
-  Repository, EntityRepository, ObjectLiteral, DeepPartial,
-} from 'typeorm';
+import * as sharp from 'sharp';
+import { Repository, EntityRepository, ObjectLiteral } from 'typeorm';
 import File from '../entities/File';
 
-const filesPath = process.env.FILES_PATH || path.join();
+const fileStoragePath = process.env.FILES_PATH || path.join(__dirname, '../../.storage/');
 
 type RawFile = {
   path?: string,
@@ -15,15 +14,24 @@ type RawFile = {
   body?: Buffer,
   headers?: ObjectLiteral,
   contentType?: string,
+  ext?: string,
 };
 
+type PreviewParams = {
+  width: number,
+  height: number,
+  left: number,
+  top: number,
+};
+
+
 @EntityRepository(File)
-export default class FileRepository<T extends File> extends Repository<File> {
-  async createFromFileIfNotExists(rawFileRequest: RawFile): Promise<File | void | null> {
-    if (!rawFileRequest) return null;
+export default class FileRepository extends Repository<File> {
+  async createFromFileIfNotExists(rawFileRequest: RawFile): Promise<File> {
+    if (!fs.existsSync(fileStoragePath)) fs.mkdirSync(fileStoragePath, { recursive: true });
 
     const fileData = rawFileRequest.path
-      ? fs.readFileSync(rawFileRequest.path, 'utf8')
+      ? fs.readFileSync(rawFileRequest.path)
       : rawFileRequest.body;
 
     if (!fileData) {
@@ -31,12 +39,11 @@ export default class FileRepository<T extends File> extends Repository<File> {
     }
 
     const md5 = createHash('md5').update(fileData).digest('hex');
-    if (rawFileRequest.path) fs.unlinkSync(rawFileRequest.path);
     const existingFile = await this.findOne({ where: { md5 } });
-    if (!_.isEmpty(existingFile)) return existingFile;
+    if (existingFile) return existingFile;
 
-    const filePath = path.join(filesPath, md5);
-    const creationData: DeepPartial<File> = {
+    const filePath = path.join(fileStoragePath, `${md5}.${rawFileRequest.ext}`);
+    const creationData: Partial<File> = {
       md5,
       name: rawFileRequest.name,
       size: Buffer.byteLength(fileData),
@@ -45,8 +52,48 @@ export default class FileRepository<T extends File> extends Repository<File> {
         || 'application/octet-stream',
       filePath,
     };
+    fs.writeFileSync(filePath, fileData);
+    return this.save(creationData);
+  }
+
+  async createPreviewFromFile(rawFileRequest: RawFile, previewParams: PreviewParams | void): Promise<File> {
+    if (!fs.existsSync(fileStoragePath)) fs.mkdirSync(fileStoragePath, { recursive: true });
+
+    const fileData = rawFileRequest.path
+      ? fs.readFileSync(rawFileRequest.path)
+      : rawFileRequest.body;
+
+    if (!fileData) {
+      throw new Error(`File.createPreviewFromFile missing fileData ${JSON.stringify(rawFileRequest, null, 2)}`);
+    }
+    const image = sharp(fileData).rotate().clone();
+    if (previewParams) image.extract(previewParams);
+    const resizeOptions: sharp.ResizeOptions = {
+      fit: 'contain',
+      background: {
+        r: 255, g: 255, b: 255, alpha: 1,
+      },
+    };
+    if (!previewParams) resizeOptions.position = 'centre';
+    const previewFileData = await image
+      .resize(800, 484, resizeOptions)
+      .jpeg()
+      .toBuffer();
+
+    const md5 = createHash('md5').update(previewFileData).digest('hex');
+    const existingFile = await this.findOne({ where: { md5 } });
+    if (existingFile) return existingFile;
+
+    const filePath = path.join(fileStoragePath, `${md5}.jpeg`);
+    const creationData: Partial<File> = {
+      md5,
+      name: rawFileRequest.name,
+      size: Buffer.byteLength(previewFileData),
+      contentType: 'image/jpeg',
+      filePath,
+    };
     fs.writeFileSync(filePath, fileData, 'utf8');
-    return this.create(creationData);
+    return this.save(creationData);
   }
 
   // destroyByIdIfNotReferenced(_id: string) {
